Fix z-value issue

+++ include/align/nonrigid_pipe.h
@@ -38,7 +38,7 @@ using PCANonRigidFittingResult = struct PCANonRigidFittingResult {
   Eigen::VectorXf expressionCoeff;
   telef::types::ImagePtrT image;
   telef::types::ImagePtrT original;
-  std::vector<uint8_t> rendered_normal;
+  std::vector<float> rendered_normal;
   std::vector<double> intensity;
   std::vector<uint16_t> rendered_depth;
   telef::types::CloudConstPtrT cloud;
